+++ CHANGELOG.md
@@ -25,6 +25,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Hero Detail remote image loading.
 - Added mock generation script.
 - Routing target.
+- Navigator component.
 
 ### Changed